refactor: move evaluation overview to a separate file
